+++ unit-2/homework/src/main/java/pro/it/sis/javacourse/IceDamage.java
@@ -1,7 +1,7 @@
 package pro.it.sis.javacourse;
 
 public class IceDamage extends Damage {
-    public IceDamage( int strength) {
+    public IceDamage(double strength) {
         super(DamageType.ice, strength);
     }
 }
