+++ unit-2/homework/src/main/java/pro/it/sis/javacourse/PhysicalDamage.java
@@ -1,7 +1,7 @@
 package pro.it.sis.javacourse;
 
 public class PhysicalDamage extends Damage {
-    public PhysicalDamage(int strength) {
+    public PhysicalDamage(double strength) {
         super(DamageType.physical, strength);
     }
 }
